them gio hang

+++ 23DH114270_MyStore/Models/Product.cs
@@ -24,7 +24,7 @@ namespace _23DH114270_MyStore.Models
         public int CategoryID { get; set; }
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
-        public decimal ProductPrice { get; set; }
+        public double ProductPrice { get; set; }
         public string ProductImage { get; set; }
     
         public virtual Category Category { get; set; }
